[Update] SingleActivity BackPress 구현
SingleActivity에 어울리는 BackPress 기능을 구현함. 이제 Activity는 완전히 껍데기가 되었으며 Fragment를 중심으로 동작하게 되었다.
