+++ presentation/src/main/java/com/mohaeyo/mohae/ui/activity/MainActivity.kt
@@ -11,28 +11,8 @@ import org.jetbrains.anko.toast
 
 class MainActivity : DaggerAppCompatActivity() {
 
-    private val backButtonSubject: Subject<Long> =
-        BehaviorSubject.createDefault(0L)
-            .toSerialized()
-
-    private val backButtonSubjectDisposable: Disposable = backButtonSubject
-        .buffer(2, 1)
-        .observeOn(AndroidSchedulers.mainThread())
-        .subscribe {
-            if (it[1] - it[0] <= 1500) finish()
-            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
-        }
-
-    override fun onBackPressed()
-            = backButtonSubject.onNext(System.currentTimeMillis())
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        backButtonSubjectDisposable.dispose()
     }
 }
