+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/MainFragment.kt
@@ -1,18 +1,24 @@
 package com.mohaeyo.mohae.ui.fragment.main
 
+import android.content.Context
 import android.graphics.drawable.Animatable
 import android.os.Bundle
 import android.view.View
+import androidx.activity.OnBackPressedCallback
 import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.fragment.findNavController
 import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.mohaeyo.mohae.R
+import com.mohaeyo.mohae.backButtonSubject
 import com.mohaeyo.mohae.base.DataBindingFragment
 import com.mohaeyo.mohae.databinding.FragmentMainBinding
 import com.mohaeyo.mohae.viewmodel.main.MainViewModel
 import com.mohaeyo.mohae.viewmodel.main.MainViewModelFactory
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
 import kotlinx.android.synthetic.main.fragment_main.*
+import org.jetbrains.anko.support.v4.toast
 import javax.inject.Inject
 
 
@@ -137,6 +143,25 @@ class MainFragment: DataBindingFragment<FragmentMainBinding>() {
         false
     }
 
+    private val backButtonSubjectDisposable: Disposable = backButtonSubject
+        .buffer(2, 1)
+        .observeOn(AndroidSchedulers.mainThread())
+        .subscribe {
+            if (it[1] - it[0] <= 1500) activity?.finish()
+            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
+        }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+
+        requireActivity().onBackPressedDispatcher.addCallback(
+            this, object: OnBackPressedCallback(true) {
+                override fun handleOnBackPressed() {
+                    backButtonSubject.onNext(System.currentTimeMillis())
+                }
+            })
+    }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.vm = viewModel
@@ -148,6 +173,11 @@ class MainFragment: DataBindingFragment<FragmentMainBinding>() {
 
     }
 
+    override fun onDestroy() {
+        super.onDestroy()
+        backButtonSubjectDisposable.dispose()
+    }
+
     private fun replaceFragment(action: Int)
             = childFragmentManager.primaryNavigationFragment!!.findNavController().navigate(action)
 
@@ -156,7 +186,6 @@ class MainFragment: DataBindingFragment<FragmentMainBinding>() {
             val avd = AnimatedVectorDrawableCompat.create(context!!, resId)
             main_background.setImageDrawable(avd)
             (main_background.drawable as Animatable).start()
-
         }
     }
 }
